moved connections to pages in multichain
