+++ ui/components/multichain/pages/connections/connections.js
