+++ ui/components/multichain/pages/connections/index.js
@@ -1,0 +1,1 @@
+export { Connections } from './connections';
